add recover old wallet logic

+++ healthcheck.sh
@@ -30,18 +30,28 @@ CreateWalletAndBackup() {
 	# If wallet don't exist
 	cd $PATH_CLIENT
 	checkWallet=`massa-client -p $PASSWORD wallet_info | grep -c "Address:"`
+
 	if ([ ! -e $PATH_CLIENT/wallet.dat ] || [ $checkWallet -lt 1 ])
 	then
-		# Generate wallet
-		cd $PATH_CLIENT
-		massa-client -p $PASSWORD wallet_generate_secret_key > /dev/null
-                sleep 10
-		if [ ! -f $PATH_HOME/wallet.dat ]; then
-                        cp wallet.dat $PATH_HOME/wallet.dat
-                        screen -dmS client bash -c 'massa-client -p '$PASSWORD''
-                        echo "[$(date +"%Y-%m-%d %H:%M:%S" --utc -d "+8 hours" )] Backup wallet.dat to $PATH_HOME" >>$PATH_HOME/healthcheck.txt
-                fi
+		# If wallet.dat already exists in $PATH_HOME
+		if [ -e $PATH_HOME/wallet.dat ]
+		then
+			# Copy wallet.dat from $PATH_HOME to $PATH_CLIENT
+			cp $PATH_HOME/wallet.dat $PATH_CLIENT/wallet.dat
+			echo "[$(date +"%Y-%m-%d %H:%M:%S" --utc -d "+8 hours" )] Copied wallet.dat from $PATH_HOME to $PATH_CLIENT" >>$PATH_HOME/healthcheck.txt
+		else
+			# Generate wallet
+			cd $PATH_CLIENT
+			massa-client -p $PASSWORD wallet_generate_secret_key > /dev/null
+                        sleep 10
+			if [ ! -f $PATH_HOME/wallet.dat ]; then
+                        	cp wallet.dat $PATH_HOME/wallet.dat
+                                screen -dmS client bash -c 'massa-client -p '$PASSWORD''
+                                echo "[$(date +"%Y-%m-%d %H:%M:%S" --utc -d "+8 hours" )] Backup wallet.dat to $PATH_HOME" >>$PATH_HOME/healthcheck.txt
+                        fi
+		fi
 	fi
+
         #backup  node_privkey.key
         if [ ! -e $PATH_HOME/node_privkey.key ]
 	then
